### 2018-12-17 2326dc6
ENH: make Group objects context managers, close at exit
