### 2018-12-17 3a3554d
MAINT: Remove deprecated rcparam: text.latex.unicode
