### 2018-12-17 5255fc4
MAINT: Remove pytz dependency
