+++ scripts/recipe/meta.yaml
@@ -1,35 +1,0 @@
-package:
-  name: wrighttools
-  version: none
-
-source:
-  path: ../../
-
-build:
-  number: 0
-  script: python setup.py install --single-version-externally-managed --record record.txt
-
-requirements:
-  build:
-    - python >=2.7
-    - setuptools
-  run:
-    - python >=2.7
-    - h5py
-    - matplotlib
-    - numpy
-    - python-dateutil
-    - pytz
-    - scipy
-
-test:
-  imports:
-    - WrightTools
-
-about:
-  home: https://github.com/wright-group/WrightTools
-  license: MIT
-  license_family: MIT
-  license_file: LICENSE
-  summary: Tools for loading, processing, and plotting multidimensional spectroscopy data.
-
