### 2018-12-18 #1131 @ygrange
Short note to say I won’t have time. Good luck!

--
Y. Grange

> Op 18 dec. 2018 om 02:50 heeft Blaise Thompson
<notifications@github.com> het volgende geschreven:
>
> @yochannah I suggest @ivergara and @jwuttke as potential reviewers
because they both mention "spectroscopy"
>
> I also suggest @ygrange and @Dybian for their interest in "workflow
development"
>
> all of these reviewers have indicated familiarity with Python
>
> unfortunately, I'm not aware of anyone who has interest in both OSS
and multidimensional spectroscopy---so I don't have any recommendations
with exact domain knowledge
>
> hope these humble suggestions are helpful in some way �
>
