### 2018-12-22 #1141 @whedon
[REVIEW]: WrightTools: a Python package for multidimensional spectroscopy
**Submitting author:** @untzag (<a href="http://orcid.org/0000-0002-3845-824X">Blaise Thompson</a>)
**Repository:** <a href="https://github.com/wright-group/WrightTools" target ="_blank">https://github.com/wright-group/WrightTools</a>
**Version:** 3.2.0
