### 2018-12-22 a9c6ea1
MAINT: Remove pytz dependency (#848)
