### 2018-12-22 2e7f721
ENH: make Group objects context managers, close at exit (#846)
