### 2018-12-22 64aa575
MAINT: Remove deprecated rcparam: text.latex.unicode (#847)
