### 2019-01-02 #1141 @daissi
@untzag everything is fine, except some minor comments:

1. README.md and homepage (http://wright.tools/) don't provide a description of the purpose of WrightTools. It would be nice to have a small introduction in both of them like in the paper (e.g. extend the sentence "Tools for loading, processing, and plotting multidimensional spectroscopy data.").

2. There is no commit from the last author John C. Wright, could you clarify what is his role in WrightTools' development?

3. CONTRIBUTORS file contains list of contributors, but it could be useful to have a link in this file to the doc how to contribute: http://wright.tools/en/stable/contributing.html

4. In the paper, you said "To our knowledge, WrightTools is the first MDS-focused toolkit to be freely available and openly licensed" but there is a list of alternatives packages with similar goal as WrightTools: http://wright.tools/en/stable/alternatives.html So, what is the difference between WrightTools and the others?

5. In some example scripts (quickstart-3, 4 and 5) from http://wright.tools/en/stable/quickstart.html#visualize-data, there are missing 'plt.show()' call at the end (just to be consistent with quickstart-1 and 2).
