### 2019-01-03 45c6a00
update readme.rst and index.rst to include statement of purpose
