### 2019-01-03 f3639c0
DOC: add link to contributing docs in CONTRIBUTORS
Addresses comment 3 from @daissi https://github.com/openjournals/joss-reviews/issues/1141#issuecomment-450850266
