### 2019-01-04 160f45d
DOC: update readme.rst and index.rst to include statement of purpose (#852)
* update readme.rst and index.rst to include statement of purpose

* DOC: update copyright year to 2019
