### 2019-01-04 3f32aae
DOC: add link to contributing docs in CONTRIBUTORS (#853)
Addresses comment 3 from @daissi https://github.com/openjournals/joss-reviews/issues/1141#issuecomment-450850266
