### 2019-01-04 ba5b7fa
DOC: add `plt.show` to quickstart examples source (#850)
No real effect for the web page as rendered, however it is consistent for those who use the 'Download Source code' link

Addresses review comment 5 by @daissi: https://github.com/openjournals/joss-reviews/issues/1141#issuecomment-450850266
