### 2019-01-04 a1a9a18
BUG: Fix prune behaviour with string parameter
closes #854
