### 2019-01-05 #1141 @daissi
> @daissi thanks so much for your comments! I'll try to respond to them here, if that's appropriate.
>
> > 1. README.md and homepage (http://wright.tools/) don't provide a description of the purpose of WrightTools. It would be nice to have a small introduction in both of them like in the paper (e.g. extend the sentence "Tools for loading, processing, and plotting multidimensional spectroscopy data.").
>
> I agree. My coauthor @darienmorrow was kind enough to make a PR to adding such a description: [wright-group/WrightTools#852](https://github.com/wright-group/WrightTools/pull/852). It has been merged.

:thumbsup:
>
> > 2. There is no commit from the last author John C. Wright, could you clarify what is his role in WrightTools' development?
>
> The developers of WrightTools are mostly graduate students in John's research group. When submitting this paper, we read the [JOSS authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) and talked among ourselves about whether John should be an author on this paper. While it's true that John has not contributed lines of code, he has involved himself significantly in ongoing conversations about how to best design software for MDS. He has been one of our most important users, and has given helpful feedback about what is confusing and what works well. Furthermore, we feel that WrightTools was shaped significantly by John's vision for high-throughput, flexible MDS instrumentation.
>
> Speaking personally, I'm just not sure how to apply the authorship guidelines in this case. I would appreciate guidance from the reviewers.
>

In this case, it is fine to have John Wright as last author of the paper. :thumbsup:

> > 3. CONTRIBUTORS file contains list of contributors, but it could be useful to have a link in this file to the doc how to contribute: http://wright.tools/en/stable/contributing.html
>
> I agree. My coauthor @ksunden was kind enough to make a PR: [wright-group/WrightTools#853](https://github.com/wright-group/WrightTools/pull/853). It has been merged.

:thumbsup:
>
> > 4. In the paper, you said "To our knowledge, WrightTools is the first MDS-focused toolkit to be freely available and openly licensed" but there is a list of alternatives packages with similar goal as WrightTools: http://wright.tools/en/stable/alternatives.html So, what is the difference between WrightTools and the others?
>
> Simply put, WrightTools focuses on multidimensional spectroscopy. The packages listed under "alternatives" all have some overlap (multidimensional self-describing datasets, a focus on spectroscopy...) but none of them focus on MDS-specific processing challenges.
>
> Perhaps the "alternatives" page could be amended to make this point clearer? I'll think on this more.
>

You can keep the name "alternatives" for this page, but maybe you can add this description in the top of the page so that readers can have a direct preview of what they can expect from the other software.

> > 5. In some example scripts (quickstart-3, 4 and 5) from http://wright.tools/en/stable/quickstart.html#visualize-data, there are missing 'plt.show()' call at the end (just to be consistent with quickstart-1 and 2).
>
> I agree. My coauthor @ksunden was kind enough to make a PR: [wright-group/WrightTools#850](https://github.com/wright-group/WrightTools/pull/850). It has been merged.

:thumbsup:
