### 2019-01-05 #1141 @ivergara
@untzag I'd say that everything looks in order, just a few comments/questions beyond what the previous reviewer said

1. The tag `hdf5` seems to me not relevant. `hdf5` is more like a detail of implementation on how to store data for future use by the same tool than something relevant for a future user. For sure it's important and relevant in the context of implementation and future accessibility of the data.

2. You claim that

> WrightTools defines a universal MDS data format

would you care to perhaps comment/expand on the `universal` character of it? I find the effort of trying to suggest a way to store the data in a more standardized way a good one. On the other hand creating a standard from small input doesn't seem universal to me. It's great that it's well documented and based on hdf5. Perhaps I'd suggest you to make an independent paper just discussing and defining this data format.

3. More of an observation, but the name `WrightTools` doesn't really evoke any kind of relationship with MDS or spectroscopy. Although it makes clear the origin of affiliation, but as such community contributions might suffer. I have the same kind of observation to the name of the data format.
