### 2019-01-08 90be949
BUG: Fix prune behaviour with string parameter (#854)
closes #854
