+++ tests/data/prune.py
@@ -47,6 +47,18 @@ def test_prune_number():
     data.close()
 
 
+def test_prune_string():
+    p = datasets.PyCMDS.wm_w2_w1_000
+    data = wt.data.from_PyCMDS(p)
+    data.create_constant("d1")
+    data.prune("pyro1")
+    assert len(data.variables) == 4
+    assert set(data.variable_names) == {"wm", "w2", "w1", "d1"}
+    assert len(data.channels) == 1
+    assert data.channel_names[0] == "pyro1"
+    data.close()
+
+
 def test_prune_tuple():
     p = datasets.PyCMDS.wm_w2_w1_000
     data = wt.data.from_PyCMDS(p)
@@ -63,4 +75,5 @@ if __name__ == "__main__":
     test_prune_default()
     test_prune_false()
     test_prune_number()
+    test_prune_string()
     test_prune_tuple()
