### 2019-01-08 7abf492
clarify meaning of alternative
