### 2019-01-08 #1141 @untzag
@ivergara thanks for the review

> 1. The tag hdf5 seems to me not relevant. hdf5 is more like a detail of implementation on how to store data for future use by the same tool than something relevant for a future user. For sure it's important and relevant in the context of implementation and future accessibility of the data.

I don't have strong feelings about the tags, so I've made a [PR](https://github.com/wright-group/WrightTools/pull/856) to remove the hdf5 tag.

> 2. You claim that WrightTools defines a universal MDS data format would you care to perhaps comment/expand on the universal character of it? I find the effort of trying to suggest a way to store the data in a more standardized way a good one. On the other hand creating a standard from small input doesn't seem universal to me. It's great that it's well documented and based on hdf5. Perhaps I'd suggest you to make an independent paper just discussing and defining this data format.

Great question. I think your critique that we currently have a relatively small number of input formats is valid. When talking about "universality", what comes to my mind is the different kinds of multidimensional experiments that the wt5 file format can describe. I do think that our format is self-evidently extensible to a wide range of experimental combinations, and in fact we do regularly use WrightTools to process dramatically different multidimensional data-sets.

I love your idea of working to more formally describe our internal file conventions. That's something we're interested in, especially if it means that other members of our community can write interfaces to our format in their preferred language (e.g. MATLAB).

> 3. More of an observation, but the name WrightTools doesn't really evoke any kind of relationship with MDS or spectroscopy. Although it makes clear the origin of affiliation, but as such community contributions might suffer. I have the same kind of observation to the name of the data format.

Yeah, if we were starting it today I don't think we would name this package WrightTools. It was a great name when we first started out with a goal of sharing some code between graduate students in the Wright Group, but now that it has grown into something we want to share with everyone the name makes less sense... We'll think more carefully next time we name something :thinking:.
