### 2019-01-09 7dcb143
BUG/ENH: Allow paths to be passed into collection constructor
Closes #535
