### 2019-01-10 23d60ef
recover contour in Quick2D (resolves #815)
