### 2019-01-10 8589245
channel mag and max
