### 2019-01-10 a808628
ENH: recover contour in Quick2D (resolves #815) (#859)
* recover contour in Quick2D (resolves #815)

* channel mag and max

* change how levels are calculated
