### 2019-01-15 #1141 @yochannah
Fantastic! Okay, @untzag  - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive?
