### 2019-01-15 883fb0c
Move to service instead of involking xvfb directly
