+++ .travis.yml
@@ -16,8 +16,6 @@ install:
   - pip install cython
   - pip install .
   - pip install -U pytest "coverage<5"
-before_script:
-- "export DISPLAY=:99.0"
-- "sh -e /etc/init.d/xvfb start"
-- sleep 3 # give xvfb some time to start
+services:
+  - xvfb
 script: python setup.py test
