### 2019-01-15 #1141 @yochannah
@untzag Thanks for this! I have a couple of minor hopefully final comments -

First - The list of paper authors in the proof above is slightly different from the list shown in the archive. We should either add them all to the paper or remove from the archive by editing the metadata.

Second, I noticed while I was doing a final scan of the manuscript that the full stop / period often comes before the citation which is a bit odd -

>Matplotlib supports one, two, and even three-dimensional plotting. (Hunter, 2007)

It would be a little bit tidier to make sure the citation is in the same sentence, i.e.

>Matplotlib supports one, two, and even three-dimensional plotting (Hunter, 2007).

This happens several times throughout the paper - would you mind updating all instances?

Once this is done, I think we'll be ready to go!
