### 2019-01-16 f64a4eb
JOSS: update punctuation to be after period (#861)
