### 2019-01-16 #1141 @untzag
@yochannah

> The list of paper authors in the proof above is slightly different from the list shown in the archive. We should either add them all to the paper or remove from the archive by editing the metadata.

In this case I think the best thing to do is to create a one-off zenodo entry for JOSS:

[10.5281/zenodo.2541776](http://doi.org/10.5281/zenodo.2541776)

Hope this works for you.

> It would be a little bit tidier to make sure the citation is in the same sentence.

Done and merged.
