### 2019-01-16 2722e0f
MAINT: Use scipy.signal instead of skimage for downscale
Removes dependency scikit-image
