### 2019-01-16 eaaacbf
TST: remove plt.show()
