### 2019-01-16 dfa8c3c
Update .travis.yml
