+++ .travis.yml
@@ -2,11 +2,11 @@ language: python
 matrix:
     include:
         - python: 3.6
-          dist: trusty
+          dist: xenial
           sudo: false
         - python: 3.7
           dist: xenial
-          sudo: true
+          sudo: false
 addons:
     apt:
         packages:
