### 2019-01-16 3bc09ce
remove sudo=false
https://blog.travis-ci.com/2018-10-04-combining-linux-infrastructures
