+++ .travis.yml
@@ -3,10 +3,8 @@ matrix:
     include:
         - python: 3.6
           dist: xenial
-          sudo: false
         - python: 3.7
           dist: xenial
-          sudo: false
 addons:
     apt:
         packages:
