### 2019-01-16 d7d9e2e
MAINT: use convolve instead of skimage for downscale
Alternative implementation to #862
