### 2019-01-16 #1141 @yochannah
@arfon Everything looks in order here - I think we're ready to accept and publish! 📖

Thanks @untzag for submitting this paper and working with us to get this ready for publication, and thank you @ivergara and @daissi for putting in time and effort to review. 💯 🥇
